Prometheus Alaerts

+++ bin/stop.sh
@@ -37,12 +37,20 @@ export DOCKER_REG=""
 export DOCKER_REG_USER=""
 export DOCKER_REG_PASSWD=""
 export DOCKER_REPO=""
+export KATANA_MONITORING=""
+export KATANA_HOST=""
+
+# Check if there are any running slices
+running_slices=$(curl -s http://localhost:8000/api/slice)
+if [[ ${running_slices} != "[]" && ${running_slices} != "" ]]; then
+    read -rp "There are running slices. Are you sure you want to stop the SM? [y/N] > " ans
+    if [[ ${ans} != "y" ]]; then
+        exit 1
+    fi
+fi
 
 # Stop the containers
 docker-compose -f ${DIR}/docker-compose.yaml down ${options}
-
-# Remove the monitoring variable
-sed -i 's/KATANA_MONITORING=.*/KATANA_MONITORING=/' "${DIR}/katana-mngr/.env"
 
 # Remove the katana-log files
 rm -f katana-nbi/katana.log* katana-mngr/katana.log*
